test: Adicionado funcionalidade para detalhar treino e exibir conteúdo do treino; fase GREEN TDD; closes #77
