+++ backend/alembic/versions/e5f6a7b8c9d0_adiciona_plan_json_a_user_workout.py
@@ -1,0 +1,25 @@
+"""adiciona plan_json a user_workout
+
+Revision ID: e5f6a7b8c9d0
+Revises: d4e5f6a7b8c9
+Create Date: 2026-06-30 12:00:00.000000
+
+"""
+from typing import Sequence, Union
+
+from alembic import op
+import sqlalchemy as sa
+
+
+revision: str = "e5f6a7b8c9d0"
+down_revision: Union[str, Sequence[str], None] = "d4e5f6a7b8c9"
+branch_labels: Union[str, Sequence[str], None] = None
+depends_on: Union[str, Sequence[str], None] = None
+
+
+def upgrade() -> None:
+    op.add_column("userworkout", sa.Column("plan_json", sa.Text(), nullable=True))
+
+
+def downgrade() -> None:
+    op.drop_column("userworkout", "plan_json")
