+++ front/__tests__/detalheTreino.test.tsx
@@ -12,6 +12,11 @@ jest.mock("@/services/api", () => ({
 
 jest.mock("expo-router", () => ({
   useLocalSearchParams: () => mockUseLocalSearchParams(),
+  useRouter: () => ({ back: jest.fn(), push: jest.fn() }),
+}));
+
+jest.mock("react-native-safe-area-context", () => ({
+  useSafeAreaInsets: () => ({ top: 0, bottom: 0, left: 0, right: 0 }),
 }));
 
 import WorkoutDetailScreen from "@/app/workout/[id]";
